seeding issue fexed (already done by chris), fortune cookies show up on main page
