+++ views/homepage.handlebars
@@ -1,0 +1,3 @@
+<h2>Your Daily Fortune: {{fortune_response}}
+    -Sun Tzu, Art Of War (not really)
+</h2>
